fix unstable test
