+++ executor/memory_test.go
@@ -21,7 +21,6 @@ import (
 	"runtime/debug"
 	"testing"
 
-	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
 )
@@ -88,29 +87,4 @@ func memDiff(m1, m2 uint64) uint64 {
 		return m1 - m2
 	}
 	return m2 - m1
-}
-
-func TestGlobalMemoryTrackerOnCleanUp(t *testing.T) {
-	// TODO: assert the memory consume has happened in another way
-	originConsume := executor.GlobalMemoryUsageTracker.BytesConsumed()
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (id int)")
-
-	// assert insert
-	tk.MustExec("insert t (id) values (1)")
-	tk.MustExec("insert t (id) values (2)")
-	tk.MustExec("insert t (id) values (3)")
-	afterConsume := executor.GlobalMemoryUsageTracker.BytesConsumed()
-	require.Equal(t, afterConsume, originConsume)
-
-	// assert update
-	tk.MustExec("update t set id = 4 where id = 1")
-	tk.MustExec("update t set id = 5 where id = 2")
-	tk.MustExec("update t set id = 6 where id = 3")
-	afterConsume = executor.GlobalMemoryUsageTracker.BytesConsumed()
-	require.Equal(t, afterConsume, originConsume)
 }
